pyproject.toml
[tool.poetry]
name = "image-text-inserter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.4.0"
pandas = "^1.3.5"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
yapf = "^0.32.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
